Add custom autocmd to enable accurate lazy-loading
Explanation for future reference:
Certain actions do not trigger Buf{Add,NewFile,ReadPost}, like
`echo something | nvim`, so in this case UI elements are not loaded.
To fix this, the User event 'VeryVeryLazy' is added that's going to
fire on VeryLazy, if the dashboard was not loaded, or when the
dashboard is unloaded.